+++ resources/lang/de/cp.php
@@ -85,12 +85,14 @@ return [
 	'covers' => 'Cookie Content Covers',
 	'covers_instructions' => 'Bitte beachte: Jedes dieser Cookie Content Covers muss in den Code der Website eingefügt werden, um zu funktionieren. Weitere Details hierzu findest du in der Dokumentation des Addons.',
 	'covers_add_row' => 'Cookie Content Cover hinzufügen',
+	'cover_name' => 'Name',
+	'cover_name_instructions' => 'Interne Bezeichnung (wird nicht auf der Website angezeigt)',
+	'cover_name_placeholder' => 'Cover für Google Maps',
 	'cover_handle' => 'Handle',
 	'cover_handle_instructions' => 'Dies ist der interne Name für die Cookie Kategorie.',
 	'cover_handle_placeholder' => 'beispiel',
 	'cover_categories' => 'Zu akzeptierende Cookie Kategorien',
 	'cover_categories_instructions' => 'Welche Cookie Kategorien müssen von den Website-Besucher_innen akzeptiert werden, damit das Cookie Content Cover auf magische Weise verschwindet?',
-	'cover_categories_placeholder' => 'essenziell,statistisch',
 	'cover_title' => 'Überschrift des Cookie Content Covers',
 	'cover_title_instructions' => 'Noch eine Überschrift (sie wird auf der Website angezeigt), die den Website-Besucher_innen ins Auge fallen. Nutze sie!',
 	'cover_title_placeholder' => 'Dieser Inhalt kann nicht angezeigt werden',
@@ -115,4 +117,8 @@ return [
 	'license_warning_toast_0' => 'Cookie Byte hat keine Lizenz.',
 	'license_warning_toast_1' => 'Kaufe eine Cookie Byte Lizenz&nbsp;<a href="https://statamic.com/addons/dryven/cookie-byte">hier</a>.',
 	'license_warning_toast_2' => 'Du musst eine Lizenz für Cookie Byte kaufen.',
+
+	/* FIELDTYPES */
+	'cookie_cover_fieldtype_title' => 'Cookie Content Cover',
+	'cookie_category_fieldtype_title' => 'Cookie Kategorie',
 ];