+++ src/Exceptions/CookieByteException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace DDM\CookieByte\Exceptions;
+
+use Exception;
+
+class CookieByteException extends Exception
+{
+}