+++ src/Exceptions/CookieCoverException.php
@@ -2,18 +2,6 @@
 
 namespace DDM\CookieByte\Exceptions;
 
-use Exception;
-
-/**
- * Class CookieCoverException
- * @package DDM\CookieByte\Exception
- * @author  dryven
- */
-class CookieCoverException extends Exception
+class CookieCoverException extends CookieByteException
 {
-
-	public function __toString()
-	{
-		parent::__toString();
-	}
 }