+++ src/Exceptions/CookieModalException.php
@@ -2,13 +2,6 @@
 
 namespace DDM\CookieByte\Exceptions;
 
-use Exception;
-
-/**
- * Class CookieModalException
- * @package DDM\CookieByte\Exception
- * @author  dryven
- */
-class CookieModalException extends Exception
+class CookieModalException extends CookieByteException
 {
 }